feat/home : revue architecture projet
