+++ src/app/components/crypto/crypto-card/crypto-card.component.html
