+++ src/app/components/crypto/crypto-list/crypto-list.component.scss
