+++ src/app/components/header/header.component.html
@@ -1,4 +1,4 @@
-<div class="header">
+<flex-div class="header">
   <flex-div alignItems="center" justifyContent="space-between">
     <!-- //TODO: conditionner une fois le store dispo -->
     <flex-div class="header__logo">Crypto</flex-div>
@@ -14,4 +14,4 @@
       <a *ngIf="isAuthenticated" class="header__links" (click)="handleDisconnectUser()">Déconnexion</a>
     </flex-div>
   </flex-div>
-</div>
+</flex-div>
