--- src/app/views/crypto/crypto-list/crypto-list.component.html
@@ -1,7 +1,0 @@
-<!-- //TODO: mettre les crypto-card ici -->
-  <flex-div class="cryptolist" alignItems="center" direction="column" *ngIf="cryptos.length > 0">
-    <flex-div class="crypto-card__wrapper" *ngFor="let crypto of cryptos">
-      <crypto-card [crypto]="crypto"></crypto-card>
-    </flex-div>
-  </flex-div>
-
