+++ src/app/views/home/home.component.scss
@@ -1,0 +1,41 @@
+.home{
+  &__wrapper {
+    margin-top: 20px;
+  }
+
+}
+
+.user{
+  &__header{
+    margin-top: 10px;
+    font-size: 1rem;
+    font-family: 'Arial', 'serif';
+  }
+
+  &__balance{
+    margin-top: 5px;
+    font-size: 2rem;
+    font-weight: bold;
+    font-family: 'Arial', 'serif';
+  }
+
+  &__gains {
+    margin-top: 5px;
+    font-size: 0.5rem;
+    font-family: 'Arial', 'serif';
+    &-balance {
+      margin-left: 10px;
+    }
+    &--positive{
+      color: green;
+    }
+    &--negative {
+      color: red;
+    }
+  }
+}
+
+.favorite__wrapper{
+  margin-top: 10px
+}
+
