+++ src/app/views/home/home.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Crypto } from "../../models/crypto";
 
 @Component({
   selector: 'app-home',
@@ -7,9 +8,33 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomeComponent implements OnInit {
 
+  favoriteCryptos: Crypto[] = [];
+  randomCryptos: Crypto[] = [];
+
   constructor() { }
 
   ngOnInit(): void {
+    this.favoriteCryptos = [{
+      id: 1,
+      name: 'Bitcoin',
+      eur_price: 29530.8,
+      img: 'https://s2.coinmarketcap.com/static/img/coins/128x128/1.png',
+      abbreviation: 'BTC',
+    }, {
+      id: 2,
+      name: 'Bitcoin',
+      eur_price: 29530.8,
+      img: 'https://s2.coinmarketcap.com/static/img/coins/128x128/1.png',
+      abbreviation: 'BTC',
+    }, {
+      id: 3,
+      name: 'Bitcoin',
+      eur_price: 29530.8,
+      img: 'https://s2.coinmarketcap.com/static/img/coins/128x128/1.png',
+      abbreviation: 'BTC',
+    }];
+
+    this.randomCryptos = this.favoriteCryptos;
   }
 
 }
