+++ src/app/views/home/home.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {RouterModule, Routes} from "@angular/router";
 import {HomeComponent} from "./home.component";
+import {WebfullstackDesignSystemModule} from "webfullstack-design-system";
+import {CryptoComponentModule} from "../../components/crypto/crypto-component.module";
 
 const routes: Routes = [
   {
@@ -14,7 +16,9 @@ const routes: Routes = [
   declarations: [HomeComponent],
   imports: [
     CommonModule,
-    RouterModule.forChild(routes)
+    RouterModule.forChild(routes),
+    WebfullstackDesignSystemModule,
+    CryptoComponentModule
   ]
 })
 export class HomeModule { }
